+++ tfjs-core/src/gradients/SpaceToBatchND_grad.ts
@@ -15,15 +15,15 @@
  * =============================================================================
  */
 
-import {BatchToSpaceND, BatchToSpaceNDAttrs} from '../kernel_names';
+import {SpaceToBatchND, SpaceToBatchNDAttrs} from '../kernel_names';
 import {GradConfig, NamedAttrMap} from '../kernel_registry';
-import {spaceToBatchND} from '../ops/space_to_batch_nd';
+import {batchToSpaceND} from '../ops/batch_to_space_nd';
 import {Tensor} from '../tensor';
 
-export const batchToSpaceNDGradConfig: GradConfig = {
-  kernelName: BatchToSpaceND,
+export const spaceToBatchNDGradConfig: GradConfig = {
+  kernelName: SpaceToBatchND,
   gradFunc: (dy: Tensor, saved: Tensor[], attrs: NamedAttrMap) => {
-    const {blockShape, crops} = attrs as {} as BatchToSpaceNDAttrs;
-    return {x: () => spaceToBatchND(dy, blockShape, crops)};
+    const {blockShape, paddings} = attrs as {} as SpaceToBatchNDAttrs;
+    return {x: () => batchToSpaceND(dy, blockShape, paddings)};
   }
 };
