+++ tfjs-core/src/public/chained_ops/register_all_chained_ops.ts
@@ -41,6 +41,7 @@ import './pad';
 import './separable_conv2d';
 import './split';
 import './squared_difference';
+import './space_to_batch_nd';
 import './sub';
 import './tile';
 import './transpose';
