+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -49,6 +49,7 @@ const CHAINED_OPS = [
   'pad',
   'max',
   'separableConv2d',
+  'spaceToBatchND',
   'split',
   'square',
   'sub',
