+++ tfjs-core/src/public/chained_ops/space_to_batch_nd.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2020 Google Inc. All Rights Reserved.
+ * Copyright 2020 Google LLC. All Rights Reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -14,16 +14,19 @@
  * limitations under the License.
  * =============================================================================
  */
+import {spaceToBatchND} from '../../ops/space_to_batch_nd';
+import {Tensor} from '../../tensor';
+import {Rank} from '../../types';
 
-import {BatchToSpaceND, BatchToSpaceNDAttrs} from '../kernel_names';
-import {GradConfig, NamedAttrMap} from '../kernel_registry';
-import {spaceToBatchND} from '../ops/space_to_batch_nd';
-import {Tensor} from '../tensor';
-
-export const batchToSpaceNDGradConfig: GradConfig = {
-  kernelName: BatchToSpaceND,
-  gradFunc: (dy: Tensor, saved: Tensor[], attrs: NamedAttrMap) => {
-    const {blockShape, crops} = attrs as {} as BatchToSpaceNDAttrs;
-    return {x: () => spaceToBatchND(dy, blockShape, crops)};
+declare module '../../tensor' {
+  interface Tensor<R extends Rank = Rank> {
+    spaceToBatchND<R extends Rank>(blockShape: number[], paddings: number[][]):
+        Tensor<R>;
   }
+}
+
+Tensor.prototype.spaceToBatchND = function<R extends Rank>(
+    blockShape: number[], paddings: number[][]): Tensor<R> {
+  this.throwIfDisposed();
+  return spaceToBatchND(this, blockShape, paddings);
 };
