+++ tfjs-core/src/register_all_gradients.ts
@@ -32,6 +32,7 @@ import {lrnGradConfig} from './gradients/LRN_grad';
 import {maxGradConfig} from './gradients/Max_grad';
 import {oneHotGradConfig} from './gradients/OneHot_grad';
 import {padV2GradConfig} from './gradients/PadV2_grad';
+import {spaceToBatchNDGradConfig} from './gradients/SpaceToBatchND_grad';
 import {splitVGradConfig} from './gradients/SplitV_grad';
 import {squareGradConfig} from './gradients/Square_grad';
 import {squaredDifferenceGradConfig} from './gradients/SquaredDifference_grad';
@@ -43,29 +44,18 @@ import {registerGradient} from './kernel_registry';
 
 // Export all kernel configs here so that the package can auto register them
 const gradConfigs: GradConfig[] = [
-  addGradConfig,
-  addNGradConfig,
-  batchMatMulGradConfig,
-  batchToSpaceNDGradConfig,
-  broadcastToGradConfig,
-  concatGradConfig,
-  conv2DGradConfig,
-  conv2DBackpropInputGradConfig,
-  conv3DGradConfig,
-  depthwiseConv2dNativeGradConfig,
-  divGradConfig,
-  fusedBatchNormGradConfig,
-  greaterEqualGradConfig,
-  identityGradConfig,
-  lrnGradConfig,
-  oneHotGradConfig,
-  padV2GradConfig,
-  splitVGradConfig,
-  maxGradConfig,
-  squareGradConfig,
-  squaredDifferenceGradConfig,
-  tileGradConfig,
-  transposeGradConfig,
+  addGradConfig,          addNGradConfig,
+  batchMatMulGradConfig,  batchToSpaceNDGradConfig,
+  broadcastToGradConfig,  concatGradConfig,
+  conv2DGradConfig,       conv2DBackpropInputGradConfig,
+  conv3DGradConfig,       depthwiseConv2dNativeGradConfig,
+  divGradConfig,          fusedBatchNormGradConfig,
+  greaterEqualGradConfig, identityGradConfig,
+  lrnGradConfig,          oneHotGradConfig,
+  padV2GradConfig,        splitVGradConfig,
+  maxGradConfig,          spaceToBatchNDGradConfig,
+  squareGradConfig,       squaredDifferenceGradConfig,
+  tileGradConfig,         transposeGradConfig,
   subGradConfig
 ];
