+++ tfjs-core/src/tensor.ts
@@ -295,8 +295,6 @@ export interface OpHandler {
       strides?: [number, number]|number): T;
   unsortedSegmentSum<T extends Tensor>(
       x: T, segmentIds: Tensor1D|TensorLike1D, numSegments: number): T;
-  spaceToBatchND<T extends Tensor>(
-      x: T, blockShape: number[], paddings: number[][]): T;
   topk<T extends Tensor>(x: T, k: number, sorted: boolean):
       {values: T, indices: T};
   stridedSlice(
@@ -1180,12 +1178,6 @@ export class Tensor<R extends Rank = Rank> {
       this: T, segmentIds: Tensor1D|TensorLike1D, numSegments: number): T {
     this.throwIfDisposed();
     return opHandler.unsortedSegmentSum(this, segmentIds, numSegments);
-  }
-
-  spaceToBatchND<T extends Tensor>(
-      this: T, blockShape: number[], paddings: number[][]): T {
-    this.throwIfDisposed();
-    return opHandler.spaceToBatchND(this, blockShape, paddings);
   }
 
   topk<T extends Tensor>(this: T, k = 1, sorted = true):
